Create module to create k8s job from bossman job

+++ bossman/protos/bossman.protobuf.v1alpha1.rs
@@ -3,7 +3,7 @@
 pub struct Options {
     /// optional
     #[prost(message, optional, tag="1")]
-    pub timeout: ::std::option::Option<i32>,
+    pub timeout: ::std::option::Option<i64>,
     /// optional
     #[prost(message, optional, tag="3")]
     pub image_pull_secrets: ::std::option::Option<::std::string::String>,
